feat refactor: move MessageStatus to enums.py
