+++ api/core/app/task_pipeline/based_generate_task_pipeline.py
@@ -19,7 +19,8 @@ from core.app.entities.task_entities import (
 from core.errors.error import QuotaExceededError
 from core.model_runtime.errors.invoke import InvokeAuthorizationError, InvokeError
 from core.moderation.output_moderation import ModerationRule, OutputModeration
-from models.model import Message, MessageStatus
+from models.model import Message
+from models.enums import MessageStatus
 
 logger = logging.getLogger(__name__)
